+++ CHANGELOG.md
@@ -9,3 +9,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 - Supported locales to the list of languages app can translate to.
+## [0.2.0] - 2021-04-26
+
+## [0.1.0] - 2021-04-23
+
+### Added
+- Export all `Categories` translations in batch
+- Export `Products` for a given category in batch - Limit of 1.600 products